operator class wm.py -> general.py
